Use the correct brush for Card backgrounds
I just noticed that in the sample code given for card pattern, the cards are using the fill for app layers instead of the fill used for cards. In the Windows UI Figma Toolkit, there is another brush called "CardBackgroundFillColorDefaultBrush" which I assume is the intended fill for cards. 

The difference between Layer and Card fills is pretty major imo, layer backgrounds are less opaque and more greyish in dark mode while the card backgrounds are brighter and more opaque, both in light and dark modes.

I've tried the "CardBackgroundFillColorDefaultBrush" brush in code and it looked exactly how the cards look in the Figma file and card surfaces used in other built-in apps like Store. The layer fill this doc currently uses for cards is not accurate.
 
I've changed the brush things (idk what they're called, not really a dev haha) next to ThemeResource to CardBackgroundFillColorDefaultBrush in the sample code provided for card pattern.